Add recursive serializer and field inspection via FieldInspector
