+++ docs/_templates/layout.html
@@ -1,0 +1,4 @@
+{% extends "!layout.html" %}
+{% block extrahead %}
+    <meta name="google-site-verification" content="saewLzcrUS1lAAgNVIikKWc3DUbFcE-TWtpyw3AW8CA" />
+{% endblock %}
