+++ src/drf_yasg/app_settings.py
@@ -5,8 +5,14 @@ from rest_framework.settings import perform_import
 SWAGGER_DEFAULTS = {
     'DEFAULT_AUTO_SCHEMA_CLASS': 'drf_yasg.inspectors.SwaggerAutoSchema',
 
-    'DEFAULT_SERIALIZER_INSPECTORS': [
-        'drf_yasg.inspectors.ReferencingSerializerInspector'
+    'DEFAULT_FIELD_INSPECTORS': [
+        'drf_yasg.inspectors.ReferencingSerializerInspector',
+        'drf_yasg.inspectors.RelatedFieldInspector',
+        'drf_yasg.inspectors.ChoiceFieldInspector',
+        'drf_yasg.inspectors.FileFieldInspector',
+        'drf_yasg.inspectors.DictFieldInspector',
+        'drf_yasg.inspectors.SimpleFieldInspector',
+        'drf_yasg.inspectors.StringDefaultFieldInspector',
     ],
     'DEFAULT_FILTER_INSPECTORS': [
         'drf_yasg.inspectors.CoreAPICompatInspector',
@@ -44,7 +50,7 @@ REDOC_DEFAULTS = {
 
 IMPORT_STRINGS = [
     'DEFAULT_AUTO_SCHEMA_CLASS',
-    'DEFAULT_SERIALIZER_INSPECTORS',
+    'DEFAULT_FIELD_INSPECTORS',
     'DEFAULT_FILTER_INSPECTORS',
     'DEFAULT_PAGINATOR_INSPECTORS',
 ]
