+++ src/drf_yasg/utils.py
@@ -1,17 +1,8 @@
 import logging
 from collections import OrderedDict
 
-from django.core.validators import RegexValidator
-from django.db import models
-from django.utils.encoding import force_text
-from rest_framework import serializers, status
+from rest_framework import status
 from rest_framework.mixins import RetrieveModelMixin, DestroyModelMixin, UpdateModelMixin
-from rest_framework.schemas.inspectors import get_pk_description
-from rest_framework.settings import api_settings
-from rest_framework.utils import json, encoders
-
-from . import openapi
-from .errors import SwaggerGenerationError
 
 logger = logging.getLogger(__name__)
 
@@ -19,51 +10,9 @@ logger = logging.getLogger(__name__)
 no_body = object()
 
 
-def is_list_view(path, method, view):
-    """Check if the given path/method appears to represent a list view (as opposed to a detail/instance view).
-
-    :param str path: view path
-    :param str method: http method
-    :param APIView view: target view
-    :rtype: bool
-    """
-    # for ViewSets, it could be the default 'list' action, or a list_route
-    action = getattr(view, 'action', '')
-    method = getattr(view, action, None)
-    detail = getattr(method, 'detail', None)
-    suffix = getattr(view, 'suffix', None)
-    if action in ('list', 'create') or detail is False or suffix == 'List':
-        return True
-
-    if action in ('retrieve', 'update', 'partial_update', 'destroy') or detail is True or suffix == 'Instance':
-        # a detail_route is surely not a list route
-        return False
-
-    # for APIView, if it's a detail view it can't also be a list view
-    if isinstance(view, (RetrieveModelMixin, UpdateModelMixin, DestroyModelMixin)):
-        return False
-
-    # if the last component in the path is parameterized it's probably not a list view
-    path_components = path.strip('/').split('/')
-    if path_components and '{' in path_components[-1]:
-        return False
-
-    # otherwise assume it's a list view
-    return True
-
-
-def guess_response_status(method):
-    if method == 'post':
-        return status.HTTP_201_CREATED
-    elif method == 'delete':
-        return status.HTTP_204_NO_CONTENT
-    else:
-        return status.HTTP_200_OK
-
-
 def swagger_auto_schema(method=None, methods=None, auto_schema=None, request_body=None, query_serializer=None,
                         manual_parameters=None, operation_id=None, operation_description=None, responses=None,
-                        serializer_inspectors=None, filter_inspectors=None, paginator_inspectors=None,
+                        field_inspectors=None, filter_inspectors=None, paginator_inspectors=None,
                         **extra_overrides):
     """Decorate a view method to customize the :class:`.Operation` object generated from it.
 
@@ -74,7 +23,7 @@ def swagger_auto_schema(method=None, methods=None, auto_schema=None, request_bod
 
     :param str method: for multi-method views, the http method the options should apply to
     :param list[str] methods: for multi-method views, the http methods the options should apply to
-    :param .SwaggerAutoSchema auto_schema: custom class to use for generating the Operation object
+    :param .inspectors.SwaggerAutoSchema auto_schema: custom class to use for generating the Operation object
     :param .Schema,.SchemaRef,.Serializer request_body: custom request body, or :data:`.no_body`. The value given here
         will be used as the ``schema`` property of a :class:`.Parameter` with ``in: 'body'``.
 
@@ -116,14 +65,14 @@ def swagger_auto_schema(method=None, methods=None, auto_schema=None, request_bod
         * a :class:`.Response` object will be used as-is; however if its ``schema`` attribute is a ``Serializer``,
           it will automatically be converted into a :class:`.Schema`
 
-    :param list[.SerializerInspector] serializer_inspectors: extra serializer inspectors;
-        these will be tried before :attr:`.serializer_inspectors` on the :class:`.SwaggerAutoSchema` instance
-    :param list[.FilterInspector] filter_inspectors: extra filter inspectors;
-        these will be tried before :attr:`.filter_inspectors` on the :class:`.SwaggerAutoSchema` instance
-    :param list[.PaginatorInspector] paginator_inspectors: extra paginator inspectors;
-        these will be tried before :attr:`.paginator_inspectors` on the :class:`.SwaggerAutoSchema` instance
+    :param list[.FieldInspector] field_inspectors: extra serializer and field inspectors; these will be tried
+        before :attr:`.ViewInspector.field_inspectors` on the :class:`.inspectors.SwaggerAutoSchema` instance
+    :param list[.FilterInspector] filter_inspectors: extra filter inspectors; these will be tried before
+        :attr:`.ViewInspector.filter_inspectors` on the :class:`.inspectors.SwaggerAutoSchema` instance
+    :param list[.PaginatorInspector] paginator_inspectors: extra paginator inspectors; these will be tried before
+        :attr:`.ViewInspector.paginator_inspectors` on the :class:`.inspectors.SwaggerAutoSchema` instance
     :param extra_overrides: extra values that will be saved into the ``overrides`` dict; these values will be available
-        in the handling :class:`.SwaggerAutoSchema` instance via ``self.overrides``
+        in the handling :class:`.inspectors.SwaggerAutoSchema` instance via ``self.overrides``
     """
 
     def decorator(view_method):
@@ -137,7 +86,7 @@ def swagger_auto_schema(method=None, methods=None, auto_schema=None, request_bod
             'responses': responses,
             'filter_inspectors': list(filter_inspectors) if filter_inspectors else None,
             'paginator_inspectors': list(paginator_inspectors) if filter_inspectors else None,
-            'serializer_inspectors': list(serializer_inspectors) if serializer_inspectors else None,
+            'field_inspectors': list(field_inspectors) if field_inspectors else None,
         }
         data = {k: v for k, v in data.items() if v is not None}
         data.update(extra_overrides)
@@ -188,319 +137,46 @@ def swagger_auto_schema(method=None, methods=None, auto_schema=None, request_bod
     return decorator
 
 
-def get_model_field(queryset, field_name):
-    """Try to get information about a model and model field from a queryset.
+def is_list_view(path, method, view):
+    """Check if the given path/method appears to represent a list view (as opposed to a detail/instance view).
 
-    :param queryset: the queryset
-    :param field_name: the target field name
-    :returns: the model and target field from the queryset as a 2-tuple; both elements can be ``None``
-    :rtype: tuple
+    :param str path: view path
+    :param str method: http method
+    :param APIView view: target view
+    :rtype: bool
     """
-    model = getattr(queryset, 'model', None)
-    try:
-        model_field = model._meta.get_field(field_name)
-    except Exception:  # pragma: no cover
-        model_field = None
+    # for ViewSets, it could be the default 'list' action, or a list_route
+    action = getattr(view, 'action', '')
+    method = getattr(view, action, None)
+    detail = getattr(method, 'detail', None)
+    suffix = getattr(view, 'suffix', None)
+    if action in ('list', 'create') or detail is False or suffix == 'List':
+        return True
 
-    return model, model_field
+    if action in ('retrieve', 'update', 'partial_update', 'destroy') or detail is True or suffix == 'Instance':
+        # a detail_route is surely not a list route
+        return False
 
+    # for APIView, if it's a detail view it can't also be a list view
+    if isinstance(view, (RetrieveModelMixin, UpdateModelMixin, DestroyModelMixin)):
+        return False
 
-model_field_to_swagger_type = [
-    (models.AutoField, (openapi.TYPE_INTEGER, None)),
-    (models.BinaryField, (openapi.TYPE_STRING, openapi.FORMAT_BINARY)),
-    (models.BooleanField, (openapi.TYPE_BOOLEAN, None)),
-    (models.NullBooleanField, (openapi.TYPE_BOOLEAN, None)),
-    (models.DateTimeField, (openapi.TYPE_STRING, openapi.FORMAT_DATETIME)),
-    (models.DateField, (openapi.TYPE_STRING, openapi.FORMAT_DATE)),
-    (models.DecimalField, (openapi.TYPE_NUMBER, None)),
-    (models.DurationField, (openapi.TYPE_INTEGER, None)),
-    (models.FloatField, (openapi.TYPE_NUMBER, None)),
-    (models.IntegerField, (openapi.TYPE_INTEGER, None)),
-    (models.IPAddressField, (openapi.TYPE_STRING, openapi.FORMAT_IPV4)),
-    (models.GenericIPAddressField, (openapi.TYPE_STRING, openapi.FORMAT_IPV6)),
-    (models.SlugField, (openapi.TYPE_STRING, openapi.FORMAT_SLUG)),
-    (models.TextField, (openapi.TYPE_STRING, None)),
-    (models.TimeField, (openapi.TYPE_STRING, None)),
-    (models.UUIDField, (openapi.TYPE_STRING, openapi.FORMAT_UUID)),
-    (models.CharField, (openapi.TYPE_STRING, None)),
-]
+    # if the last component in the path is parameterized it's probably not a list view
+    path_components = path.strip('/').split('/')
+    if path_components and '{' in path_components[-1]:
+        return False
+
+    # otherwise assume it's a list view
+    return True
 
 
-def inspect_model_field(model, model_field):
-    """Extract information from a django model field instance.
-
-    :param model: the django model
-    :param model_field: a field on the model
-    :return: description, type, format and pattern extracted from the model field
-    :rtype: OrderedDict
-    """
-    if model is not None and model_field is not None:
-        for model_field_class, tf in model_field_to_swagger_type:
-            if isinstance(model_field, model_field_class):
-                swagger_type, format = tf
-                break
-        else:  # pragma: no cover
-            swagger_type, format = None, None
-
-        if format is None or format == openapi.FORMAT_SLUG:
-            pattern = find_regex(model_field)
-        else:
-            pattern = None
-
-        if model_field.help_text:
-            description = force_text(model_field.help_text)
-        elif model_field.primary_key:
-            description = get_pk_description(model, model_field)
-        else:
-            description = None
+def guess_response_status(method):
+    if method == 'post':
+        return status.HTTP_201_CREATED
+    elif method == 'delete':
+        return status.HTTP_204_NO_CONTENT
     else:
-        description = None
-        swagger_type = None
-        format = None
-        pattern = None
-
-    result = OrderedDict([
-        ('description', description),
-        ('type', swagger_type or openapi.TYPE_STRING),
-        ('format', format),
-        ('pattern', pattern)
-    ])
-    # TODO: filter none
-    return result
-
-
-def serializer_field_to_swagger(field, swagger_object_type, definitions=None, **kwargs):
-    """Convert a drf Serializer or Field instance into a Swagger object.
-
-    :param rest_framework.serializers.Field field: the source field
-    :param type[openapi.SwaggerDict] swagger_object_type: should be one of Schema, Parameter, Items
-    :param .ReferenceResolver definitions: optional, used to serialize Schemas by reference
-    :param kwargs: extra attributes for constructing the object;
-       if swagger_object_type is Parameter, ``name`` and ``in_`` should be provided
-    :return: the swagger object
-    :rtype: openapi.Parameter, openapi.Items, openapi.Schema
-    """
-    assert swagger_object_type in (openapi.Schema, openapi.Parameter, openapi.Items)
-    assert not isinstance(field, openapi.SwaggerDict), "passed field is already a SwaggerDict object"
-    title = force_text(field.label) if field.label else None
-    title = title if swagger_object_type == openapi.Schema else None  # only Schema has title
-    title = None
-    description = force_text(field.help_text) if field.help_text else None
-    description = description if swagger_object_type != openapi.Items else None  # Items has no description either
-
-    def SwaggerType(existing_object=None, **instance_kwargs):
-        if swagger_object_type == openapi.Parameter and 'required' not in instance_kwargs:
-            instance_kwargs['required'] = field.required
-        if swagger_object_type != openapi.Items and 'default' not in instance_kwargs:
-            default = getattr(field, 'default', serializers.empty)
-            if default is not serializers.empty:
-                if callable(default):
-                    try:
-                        if hasattr(default, 'set_context'):
-                            default.set_context(field)
-                        default = default()
-                    except Exception:  # pragma: no cover
-                        logger.warning("default for %s is callable but it raised an exception when "
-                                       "called; 'default' field will not be added to schema", field, exc_info=True)
-                        default = None
-
-                if default is not None:
-                    try:
-                        default = field.to_representation(default)
-                        # JSON roundtrip ensures that the value is valid JSON;
-                        # for example, sets get transformed into lists
-                        default = json.loads(json.dumps(default, cls=encoders.JSONEncoder))
-                    except Exception:  # pragma: no cover
-                        logger.warning("'default' on schema for %s will not be set because "
-                                       "to_representation raised an exception", field, exc_info=True)
-                        default = None
-
-                if default is not None:
-                    instance_kwargs['default'] = default
-
-        if swagger_object_type == openapi.Schema and 'read_only' not in instance_kwargs:
-            if field.read_only:
-                instance_kwargs['read_only'] = True
-        instance_kwargs.update(kwargs)
-        instance_kwargs.pop('title', None)
-        instance_kwargs.pop('description', None)
-
-        if existing_object is not None:
-            existing_object.title = title
-            existing_object.description = description
-            for attr, val in instance_kwargs.items():
-                setattr(existing_object, attr, val)
-            return existing_object
-
-        return swagger_object_type(title=title, description=description, **instance_kwargs)
-
-    # arrays in Schema have Schema elements, arrays in Parameter and Items have Items elements
-    ChildSwaggerType = openapi.Schema if swagger_object_type == openapi.Schema else openapi.Items
-
-    # ------ NESTED
-    if isinstance(field, (serializers.ListSerializer, serializers.ListField)):
-        child_schema = serializer_field_to_swagger(field.child, ChildSwaggerType, definitions)
-        return SwaggerType(
-            type=openapi.TYPE_ARRAY,
-            items=child_schema,
-        )
-    elif isinstance(field, serializers.Serializer):
-        if swagger_object_type != openapi.Schema:
-            raise SwaggerGenerationError("cannot instantiate nested serializer as " + swagger_object_type.__name__)
-
-        serializer = field
-        if hasattr(serializer, '__ref_name__'):
-            ref_name = serializer.__ref_name__
-        else:
-            ref_name = type(serializer).__name__
-            if ref_name.endswith('Serializer'):
-                ref_name = ref_name[:-len('Serializer')]
-
-        def make_schema_definition():
-            properties = OrderedDict()
-            required = []
-            for key, value in serializer.fields.items():
-                properties[key] = serializer_field_to_swagger(value, ChildSwaggerType, definitions)
-                if value.required:
-                    required.append(key)
-
-            return SwaggerType(
-                type=openapi.TYPE_OBJECT,
-                properties=properties,
-                required=required or None,
-            )
-
-        if not ref_name or definitions is None:
-            return make_schema_definition()
-
-        definitions.setdefault(ref_name, make_schema_definition)
-        return openapi.SchemaRef(definitions, ref_name)
-    elif isinstance(field, serializers.ManyRelatedField):
-        child_schema = serializer_field_to_swagger(field.child_relation, ChildSwaggerType, definitions)
-        return SwaggerType(
-            type=openapi.TYPE_ARRAY,
-            items=child_schema,
-            unique_items=True,  # is this OK?
-        )
-    elif isinstance(field, serializers.PrimaryKeyRelatedField):
-        if field.pk_field:
-            result = serializer_field_to_swagger(field.pk_field, swagger_object_type, definitions, **kwargs)
-            return SwaggerType(existing_object=result)
-
-        attrs = {'type': openapi.TYPE_STRING}
-        try:
-            model = field.queryset.model
-            pk_field = model._meta.pk
-        except Exception:  # pragma: no cover
-            logger.warning("an exception was raised when attempting to extract the primary key related to %s; "
-                           "falling back to plain string" % field, exc_info=True)
-        else:
-            attrs.update(inspect_model_field(model, pk_field))
-
-        return SwaggerType(**attrs)
-    elif isinstance(field, serializers.HyperlinkedRelatedField):
-        return SwaggerType(type=openapi.TYPE_STRING, format=openapi.FORMAT_URI)
-    elif isinstance(field, serializers.SlugRelatedField):
-        model, model_field = get_model_field(field.queryset, field.slug_field)
-        attrs = inspect_model_field(model, model_field)
-        return SwaggerType(**attrs)
-    elif isinstance(field, serializers.RelatedField):
-        return SwaggerType(type=openapi.TYPE_STRING)
-    # ------ CHOICES
-    elif isinstance(field, serializers.MultipleChoiceField):
-        return SwaggerType(
-            type=openapi.TYPE_ARRAY,
-            items=ChildSwaggerType(
-                type=openapi.TYPE_STRING,
-                enum=list(field.choices.keys())
-            )
-        )
-    elif isinstance(field, serializers.ChoiceField):
-        return SwaggerType(type=openapi.TYPE_STRING, enum=list(field.choices.keys()))
-    # ------ BOOL
-    elif isinstance(field, (serializers.BooleanField, serializers.NullBooleanField)):
-        return SwaggerType(type=openapi.TYPE_BOOLEAN)
-    # ------ NUMERIC
-    elif isinstance(field, (serializers.DecimalField, serializers.FloatField)):
-        # TODO: min_value max_value
-        return SwaggerType(type=openapi.TYPE_NUMBER)
-    elif isinstance(field, serializers.IntegerField):
-        # TODO: min_value max_value
-        return SwaggerType(type=openapi.TYPE_INTEGER)
-    elif isinstance(field, serializers.DurationField):
-        return SwaggerType(type=openapi.TYPE_INTEGER)
-    # ------ STRING
-    elif isinstance(field, serializers.EmailField):
-        return SwaggerType(type=openapi.TYPE_STRING, format=openapi.FORMAT_EMAIL)
-    elif isinstance(field, serializers.RegexField):
-        return SwaggerType(type=openapi.TYPE_STRING, pattern=find_regex(field))
-    elif isinstance(field, serializers.SlugField):
-        return SwaggerType(type=openapi.TYPE_STRING, format=openapi.FORMAT_SLUG, pattern=find_regex(field))
-    elif isinstance(field, serializers.URLField):
-        return SwaggerType(type=openapi.TYPE_STRING, format=openapi.FORMAT_URI)
-    elif isinstance(field, serializers.IPAddressField):
-        format = {'ipv4': openapi.FORMAT_IPV4, 'ipv6': openapi.FORMAT_IPV6}.get(field.protocol, None)
-        return SwaggerType(type=openapi.TYPE_STRING, format=format)
-    elif isinstance(field, serializers.CharField):
-        # TODO: min_length max_length (for all CharField subclasses above too)
-        return SwaggerType(type=openapi.TYPE_STRING)
-    elif isinstance(field, serializers.UUIDField):
-        return SwaggerType(type=openapi.TYPE_STRING, format=openapi.FORMAT_UUID)
-    # ------ DATE & TIME
-    elif isinstance(field, serializers.DateField):
-        return SwaggerType(type=openapi.TYPE_STRING, format=openapi.FORMAT_DATE)
-    elif isinstance(field, serializers.DateTimeField):
-        return SwaggerType(type=openapi.TYPE_STRING, format=openapi.FORMAT_DATETIME)
-    # ------ OTHERS
-    elif isinstance(field, serializers.FileField):
-        # swagger 2.0 does not support specifics about file fields, so ImageFile gets no special treatment
-        # OpenAPI 3.0 does support it, so a future implementation could handle this better
-        err = SwaggerGenerationError("FileField is supported only in a formData Parameter or response Schema")
-        if swagger_object_type == openapi.Schema:
-            # FileField.to_representation returns URL or file name
-            result = SwaggerType(type=openapi.TYPE_STRING, read_only=True)
-            if getattr(field, 'use_url', api_settings.UPLOADED_FILES_USE_URL):
-                result.format = openapi.FORMAT_URI
-            return result
-        elif swagger_object_type == openapi.Parameter:
-            param = SwaggerType(type=openapi.TYPE_FILE)
-            if param['in'] != openapi.IN_FORM:
-                raise err  # pragma: no cover
-            return param
-        else:
-            raise err  # pragma: no cover
-    elif isinstance(field, serializers.DictField) and swagger_object_type == openapi.Schema:
-        child_schema = serializer_field_to_swagger(field.child, ChildSwaggerType, definitions)
-        return SwaggerType(
-            type=openapi.TYPE_OBJECT,
-            additional_properties=child_schema
-        )
-    elif isinstance(field, serializers.ModelField):
-        return SwaggerType(type=openapi.TYPE_STRING)
-
-    # TODO unhandled fields: TimeField HiddenField JSONField
-
-    # everything else gets string by default
-    return SwaggerType(type=openapi.TYPE_STRING)
-
-
-def find_regex(regex_field):
-    """Given a ``Field``, look for a ``RegexValidator`` and try to extract its pattern and return it as a string.
-
-    :param serializers.Field regex_field: the field instance
-    :return: the extracted pattern, or ``None``
-    :rtype: str
-    """
-    regex_validator = None
-    for validator in regex_field.validators:
-        if isinstance(validator, RegexValidator):
-            if regex_validator is not None:
-                # bail if multiple validators are found - no obvious way to choose
-                return None  # pragma: no cover
-            regex_validator = validator
-
-    # regex_validator.regex should be a compiled re object...
-    return getattr(getattr(regex_validator, 'regex', None), 'pattern', None)
+        return status.HTTP_200_OK
 
 
 def param_list_to_odict(parameters):
@@ -516,3 +192,21 @@ def param_list_to_odict(parameters):
     result = OrderedDict(((param.name, param.in_), param) for param in parameters)
     assert len(result) == len(parameters), "duplicate Parameters found"
     return result
+
+
+def filter_none(obj):
+    """Remove ``None`` values from tuples, lists or dictionaries. Return other objects as-is.
+
+    :param obj:
+    :return: collection with ``None`` values removed
+    """
+    if obj is None:
+        return None
+    new_obj = None
+    if isinstance(obj, dict):
+        new_obj = type(obj)((k, v) for k, v in obj.items() if k is not None and v is not None)
+    if isinstance(obj, (list, tuple)):
+        new_obj = type(obj)(v for v in obj if v is not None)
+    if new_obj is not None and len(new_obj) != len(obj):
+        return new_obj  # pragma: no cover
+    return obj
