+++ testproj/testproj/settings.py
@@ -128,3 +128,47 @@ USE_TZ = True
 STATIC_URL = '/static/'
 
 TEST_RUNNER = 'testproj.runner.PytestTestRunner'
+
+LOGGING = {
+    'version': 1,
+    'disable_existing_loggers': True,
+    'formatters': {
+        'pipe_separated': {
+            'format': '%(asctime)s | %(levelname)s | %(name)s | %(message)s'
+        }
+    },
+    'handlers': {
+        'console_log': {
+            'level': 'DEBUG',
+            'class': 'logging.StreamHandler',
+            'stream': 'ext://sys.stdout',
+            'formatter': 'pipe_separated',
+        },
+    },
+    'loggers': {
+        'drf_yasg': {
+            'handlers': ['console_log'],
+            'level': 'DEBUG',
+            'propagate': True,
+        },
+        'django': {
+            'handlers': ['console_log'],
+            'level': 'DEBUG',
+            'propagate': True,
+        },
+        'django.db.backends': {
+            'handlers': ['console_log'],
+            'level': 'INFO',
+            'propagate': True,
+        },
+        'django.template': {
+            'handlers': ['console_log'],
+            'level': 'INFO',
+            'propagate': True,
+        },
+    },
+    'root': {
+        'handlers': ['console_log'],
+        'level': 'INFO',
+    }
+}
